feat: serialize CRDT messages (#2323)

+++ unity-renderer/Assets/Scripts/MainScripts/DCL/CRDTProtocol/CRDTMessage.cs
@@ -3,7 +3,7 @@ namespace DCL.CRDT
     public class CRDTMessage
     {
         public long key;
-        public double timestamp;
+        public long timestamp;
         public object data;
     }
 }
